requirements.txt
langchain 
python-pptx
dotenv
langchain_groq
langchain_tavily